feat: modularise block, add mocha and test suite utils
License: MIT
